+++ src/service-app/classes/CpuWorker.ts
@@ -35,13 +35,14 @@ export class CpuWorker extends DaemonWorker {
      * Skip writing energy performance preference if flag is set
      */
     private noEPPWriteQuirk: boolean;
+    private device: TUXEDODevice;
 
     constructor(tccd: TuxedoControlCenterDaemon) {
         super(10000, tccd);
         this.cpuCtrl = new CpuController(this.basePath);
 
-        const dev = this.tccd.identifyDevice();
-        if ([TUXEDODevice.SIRIUS1602, TUXEDODevice.STELLSL15A06].includes(dev)) {
+        this.device = this.tccd.identifyDevice();
+        if ([TUXEDODevice.SIRIUS1602, TUXEDODevice.STELLSL15A06].includes(this.device)) {
             this.noEPPWriteQuirk = true;
         } else {
             this.noEPPWriteQuirk = false;
@@ -166,7 +167,12 @@ export class CpuWorker extends DaemonWorker {
 
                 this.cpuCtrl.setGovernor(profile.cpu.governor);
                 if (!this.noEPPWriteQuirk) {
-                    this.cpuCtrl.setEnergyPerformancePreference(profile.cpu.energyPerformancePreference);
+                    if (this.device === TUXEDODevice.GEMINI17I04) {
+                        // Quirk for Gemini Gen4 Intel, needs EPP = performance to allow full frequency range
+                        this.cpuCtrl.setEnergyPerformancePreference("performance");
+                    } else {
+                        this.cpuCtrl.setEnergyPerformancePreference(profile.cpu.energyPerformancePreference);
+                    }
                 }
 
                 this.cpuCtrl.setGovernorScalingMinFrequency(profile.cpu.scalingMinFrequency);
@@ -312,7 +318,7 @@ export class CpuWorker extends DaemonWorker {
 
                 const currentPerformancePreference = core.energyPerformancePreference.readValue();
                 let performancePreferenceProfile: string;
-                if (!profile.cpu.useMaxPerfGov) {
+                if (!profile.cpu.useMaxPerfGov && this.device !== TUXEDODevice.GEMINI17I04) {
                     performancePreferenceProfile = profile.cpu.energyPerformancePreference
                 } else {
                     performancePreferenceProfile = "performance"